rename server_sent_events module to sse
