+++ bocadillo/__init__.py
@@ -4,7 +4,7 @@ from .media import Media
 from .middleware import Middleware
 from .recipes import Recipe
 from .staticfiles import static
-from .server_sent_events import server_event
+from .sse import server_event
 from .views import view
 from .websockets import WebSocket, WebSocketDisconnect
 from .response import Response
